requirements.txt
feedparser==6.0.11
aiohttp==3.9.5
aiosqlite==0.20.0
python-telegram-bot[job-queue]==21.3
google-generativeai==0.7.2
APScheduler==3.10.4
python-dotenv==1.0.1
